* Fixed warnings by either disabling them or fixing code. ** Disabled security warnings on string functions in particular * Added todo for line in SPI.h: #define SPI_CLOCK_DIV64 0x07  //TODO: What *should* this be (redefines DIV64) ** Should be either 0x02 or 0x07.  Code redefines 0x02 as 0x07, so commented out 0x02 one for moment (help?)

+++ gdemu/Print.cpp
@@ -45,7 +45,7 @@ void Print::write(const uint8_t *buffer, size_t size)
 
 void Print::print(const String &s)
 {
-  for (int i = 0; i < s.length(); i++) {
+  for (uint8_t i = 0; i < s.length(); i++) {
     write(s[i]);
   }
 }
@@ -78,7 +78,7 @@ void Print::print(unsigned int n, int base)
 void Print::print(long n, int base)
 {
   if (base == 0) {
-    write(n);
+    write((uint8_t)n);
   } else if (base == 10) {
     if (n < 0) {
       print('-');
@@ -92,7 +92,7 @@ void Print::print(long n, int base)
 
 void Print::print(unsigned long n, int base)
 {
-  if (base == 0) write(n);
+  if (base == 0) write((uint8_t)n);
   else printNumber(n, base);
 }
